multi readers, update world version, add datafixer, remove downgrade
