+++ SLIME_FORMAT
@@ -83,6 +83,7 @@ World version list:
  - 4: 1.13 world
  - 5: 1.14 world
  - 6: 1.16 world
+ - 7: 1.17 world
 
 -------------------------------------
