+++ slimeworldmanager-nms-v117-0/src/main/java/com/grinderwolf/swm/nms/v117/v1170SlimeNMS.java
@@ -84,7 +84,7 @@ public class v1170SlimeNMS implements SlimeNMS {
         }));
     }
 
-    private final byte worldVersion = 0x06;
+    private final byte worldVersion = 0x07;
 
     private boolean loadingDefaultWorlds = true; // If true, the addWorld method will not be skipped
