+++ slimeworldmanager-plugin/src/main/java/com/grinderwolf/swm/plugin/upgrade/Upgrade.java
@@ -5,5 +5,4 @@ import com.grinderwolf.swm.nms.CraftSlimeWorld;
 public interface Upgrade {
 
     void upgrade(CraftSlimeWorld world);
-    void downgrade(CraftSlimeWorld world);
 }
