+++ slimeworldmanager-plugin/src/main/java/com/grinderwolf/swm/plugin/upgrade/WorldUpgrader.java
@@ -3,6 +3,7 @@ package com.grinderwolf.swm.plugin.upgrade;
 import com.grinderwolf.swm.nms.CraftSlimeWorld;
 import com.grinderwolf.swm.plugin.SWMPlugin;
 import com.grinderwolf.swm.plugin.log.Logging;
+import com.grinderwolf.swm.plugin.upgrade.v117.v117WorldUpgrade;
 import com.grinderwolf.swm.plugin.upgrade.v1_14.v1_14WorldUpgrade;
 import com.grinderwolf.swm.plugin.upgrade.v1_16.v1_16WorldUpgrade;
 
@@ -14,9 +15,10 @@ public class WorldUpgrader {
     private static final Map<Byte, Upgrade> upgrades = new HashMap<>();
 
     static {
-        upgrades.put((byte) 0x05, new v1_14WorldUpgrade());
-        upgrades.put((byte) 0x07, new v1_16WorldUpgrade());
-        // Todo we need a 1_14_WorldUpgrade class as well for 0x06
+        upgrades.put((byte) 0x04, new v1_14WorldUpgrade());
+        // Todo we need a 1_14_WorldUpgrade class as well for 0x05
+        upgrades.put((byte) 0x06, new v1_16WorldUpgrade());
+        upgrades.put((byte) 0x07, new v117WorldUpgrade());
     }
 
     public static void upgradeWorld(CraftSlimeWorld world) {
@@ -36,20 +38,4 @@ public class WorldUpgrader {
         world.setVersion(serverVersion);
     }
 
-    public static void downgradeWorld(CraftSlimeWorld world) {
-        byte serverVersion = SWMPlugin.getInstance().getNms().getWorldVersion();
-
-        for (byte ver = world.getVersion(); ver > serverVersion; ver--) {
-            Upgrade upgrade = upgrades.get(ver);
-
-            if (upgrade == null) {
-                Logging.warning("Missing world upgrader for version " + ver + ". World will not be downgraded.");
-                continue;
-            }
-
-            upgrade.downgrade(world);
-        }
-
-        world.setVersion(serverVersion);
-    }
 }
