+++ slimeworldmanager-plugin/src/main/java/com/grinderwolf/swm/plugin/upgrade/v1_14/v1_14WorldUpgrade.java
@@ -152,72 +152,6 @@ public class v1_14WorldUpgrade implements Upgrade {
                         (profession == 4 ? (career == 2 ? "minecraft:leatherworker" : "minecraft:butcher") : (profession == 5 ? "minecraft:nitwit" : "minecraft:none")))));
     }
 
-    @Override
-    public void downgrade(CraftSlimeWorld world) {
-        for (SlimeChunk chunk : new ArrayList<>(world.getChunks().values())) {
-            // Update renamed blocks
-            for (int sectionIndex = 0; sectionIndex < chunk.getSections().length; sectionIndex++) {
-                SlimeChunkSection section = chunk.getSections()[sectionIndex];
-
-                if (section != null) {
-                    List<CompoundTag> palette = section.getPalette().getValue();
-
-                    for (int paletteIndex = 0; paletteIndex < palette.size(); paletteIndex++) {
-                        CompoundTag blockTag = palette.get(paletteIndex);
-                        String name = blockTag.getStringValue("Name").get();
-
-                        // The trapped chest tile entity type didn't exist until 1.13
-                        if (name.equals("minecraft:trapped_chest")) {
-                            updateBlockEntities(chunk, sectionIndex, paletteIndex, "minecraft:trapped_chest", "minecraft:chest");
-                        }
-
-                        String newName = newToOldMap.get(name);
-
-                        if (newName != null) {
-                            blockTag.getValue().put("Name", new StringTag("Name", newName));
-                        }
-                    }
-                }
-            }
-
-            if (chunk.getEntities() != null) {
-                for (CompoundTag entityTag : chunk.getEntities()) {
-                    String type = entityTag.getStringValue("id").get();
-
-                    switch (type) {
-                        case "minecraft:cat":
-                            // Cats are ocelots
-                            entityTag.getValue().put("id", new StringTag("id", "minecraft:ocelot"));
-                            break;
-                        case "minecraft:villager":
-                        case "minecraft:zombie_villager":
-                            // Villager data has changed
-                            CompoundTag dataTag = entityTag.getAsCompoundTag("VillagerData").get();
-                            String profession = dataTag.getStringValue("profession").get();
-                            int[] professionData = getVillagerProfession(profession);
-
-                            entityTag.getValue().remove("VillagerData");
-                            entityTag.getValue().put("Profession", new IntTag("Profession", professionData[0]));
-                            entityTag.getValue().put("Career", new IntTag("Career", professionData[1]));
-                            entityTag.getValue().put("CareerLevel", new IntTag("Career", 1));
-                            break;
-                        case "minecraft:banner":
-                            // The illager banners changed the translation message
-                            Optional<String> customName = entityTag.getStringValue("CustomName");
-
-                            if (customName.isPresent()) {
-                                String newName = customName.get().replace("\"translate\":\"block.minecraft.ominous_banner\"",
-                                        "\"translate\":\"block.minecraft.illager_banner\"");
-
-                                entityTag.getValue().put("CustomName", new StringTag("CustomName", newName));
-                            }
-                            break;
-                    }
-                }
-            }
-        }
-    }
-
     private int[] getVillagerProfession(String profession) {
         switch (profession) {
             case "minecraft:farmer":
